move unstable code to benches/bench.rs
